heictojpg/requirements.txt
Pillow>=11.2.1
pillow-heif>=1.1.1
Flask>=3.1.0
